storage: determine correctly case when dmsetup doesn't find any devices
In the commit

  commit 57153711ef716b0e76d84733aa224102c79e8022
  Author: Vojtech Juranek <vjuranek@redhat.com>
  Date:   Fri Sep 13 14:24:48 2019 +0200

we replaced deprecated misc.execCmd() with commands.run() in
devicemapper module.

misc.execCmd() returns output with splitted lines, while commands.run()
returns raw output and output lines need to be splitted in the code
calling this function. We did this in the patch, but forgot to change
appropriately branch in getPathsStatus() function in which we determine
the case when no device mapper device is found. In this case output
contains only one line with no colon ("No devices found"). Using
unsplitted output, we fail to detemine this case and fail with

  error=not enough values to unpack (expected 2, got 1) (dispatcher:87)

Use splitted output for determining the case when no device is found.

Change-Id: I3f09cc66dc64ebc8a6b68e0126bca7f63ce178b5
Bug-Url: https://bugzilla.redhat.com/1766595
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>