backup: add 'checkpoint' property for each backup disk
Add 'checkpoint' = true/false for each disk that participates
in a VM backup.
This information will guide VDSM on which disk should be included in the
checkpoint creation for incremental VM backup.

The disk will only be included in the checkpoint if the disk format is
'cow', we cannot add bitmaps to raw disks.

So now the request data from Engine to VDSM contains the following
structure:
{
    backup_id: backup-1,
    disks: [
        {
            "img_id": "disk1",,
            "domain_id": "domain1",
            "volume_id": "volume1",
            "checkpoint": true/false,
        },
        {
            "img_id": "disk2",,
            "domain_id": "domain2",
            "volume_id": "volume2",
            "checkpoint": true/false,
        },
    from_checkpoint_id: "from_checkpoint_id",
    to_checkpoint_id: "to_checkpoint_id"
}

The 'checkpoint' property will be used in the followup patch.

Change-Id: If1bc1c49f6cd1a4d4b521b54753c0fd08ca595e3
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>