virt: async snapshot recovery
This patch will let the vdsm to recovery the snapshot job.
If the VDSM is restarted while performing a snapshot,
a new recovery job will start.
The recovery will have the original job id and configuration.
It will monitor libvirt by checking the snapshot jobstatus.
After the snapshot operation finishes, the recovery will
perform teardown.

VDSM sends a command to libvirt in order to execute
the snapshot. In case of recovery, the job might be
running in libvirt. In domjobinfo we can see what is
the job's status. Only one job can be executed at a specific
time per VM. Therefore, using our VM metadata and checking
libvirt domjobinfo status will give us good identification
to the job status.

In case it succeeds, we will have the new snapshot ready
and the job will reported as done successfully to the engine.
If not, it will be reported as failed.

Change-Id: I3b85b401c4a0de1736cf420a55bc49b376949ec6
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>