virt: Generalized support for hot unplugged device teardown
We receive a device removal event from libvirt once a device hot
unplug is finished.  Since device hot unplug can take an arbitrary
amount of time, this is an asynchronous event and we handle it in a
callback.

At the time this event arrives, the device is already gone and is no
longer present in the domain XML.  We get only the alias of the
removed device and nothing more.  In order to perform teardown actions
on the device, we need to store the hot unplugged device somewhere.
Device instances are currently used for the purpose, but we are trying
to get rid of them, so we need another mechanism to track the hot
unplugged devices.

This patch adds a simplified device class that can be used (in future
patches) to handle hot plugged devices without tracking them in
Vm._devices and extends device removal callback with looking up the
(simplified) removed device in a dictionary of devices being currently
unplugged.

The device being unplugged is lost on Vdsm restart.  This is not a
fundamental regression, since the same could happen if the device was
removed when Vdsm wasn't running and thus has no longer been present
in the domain XML on new Vdsm restart.  Device teardown is not
performed in both the cases.  Vdsm shouldn't be restarted during
device hot unplug.

Change-Id: Ib027c02d64954c6014877df928eee723c8dc82bc
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>