net, nmstate: Support custom bridge options
The custom bridge options have been supported by the
initscripts (ifcfg).

This change introduces support for configuring bridge options
through sysfs directly.
The bridge options are persisted by VDSM (through unified persistence)
and reported by the network caps.

The new Bridge class does not support the creation of the bridge,
because it is not needed for now. Once is necessary, an addition
of "create()" is available.

Change-Id: Id43b4ddde90ebf58cd44649afd934c7a3508a377
Bug-Url: https://bugzilla.redhat.com/1790503
Signed-off-by: Bell Levin <blevin@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>