net: Fix '\D' py3 escape deprecation
Runnning the unit test, we get the following warning:
/vdsm/lib/vdsm/network/models.py:442:
DeprecationWarning: invalid escape sequence \D
    nicsRexp = re.compile("^(\D*)(\d*)(.*)$")

This warning will be a syntax error in the future and has to
be changed.

Strings in py3 are treated as unicode, and thus escaping the
following - '\D'- results in escaping the character 'd'.
To avoid this faulty escape we should convert the string
to a raw string, which uses different rules for backslashes.

Change-Id: I35c01e9d445918ebf71342edd6069382fb264689
Signed-off-by: Bell Levin <blevin@redhat.com>