net: Fix '\w' py3 escape deprecation
Runnning the unit test, we get the following warnings:
/home/blevin/projects/vdsm/lib/vdsm/network/link/validator.py:38
  /home/blevin/projects/vdsm/lib/vdsm/network/link/validator.py:38:
DeprecationWarning: invalid escape sequence \w
    bond for bond in bonds if not re.match('^bond\w+$', bond)

/home/blevin/projects/vdsm/lib/vdsm/network/link/validator.py:44
  /home/blevin/projects/vdsm/lib/vdsm/network/link/validator.py:44:
DeprecationWarning: invalid escape sequence \w
    and not re.match('^bond\w+$', net_attrs['bonding'])

This warning will be a syntax error in the future and has to
be changed.

Strings in py3 are treated as unicode, and thus escaping the
following - '\w'- results in escaping the character 'w'.
To avoid this faulty escape we should convert the string
to a raw string, which uses different rules for backslashes.

Change-Id: I1cfe3aebbfc2033b075f994853592c7d8562500e
Signed-off-by: Bell Levin <blevin@redhat.com>