net, nmstate, tests: Add test for moving dynamic address
On IPv4, the client-id could be defined with the mac
address of the interface as its only source. Since Nm 1.22
the same parameter can be specified also for IPv6.
Using the mac address as the only client ID source allows
moving a dynamic address from a nic to a parent interface
(bond, bridge, etc...).

Change-Id: Ie046d9b2dd6a749526cb3604d0f8f77c3f44e0da
Signed-off-by: Ales Musil <amusil@redhat.com>