virt: Update guest agent API version before making migration params
When a VM is migrated, guestAgentAPIVersion metadata may not be
present on the destination.  Then the destination assumes the API
version is 0 and after_migration event is not sent to the guest.

The problem is that metadata is updated with guestAgentAPIVersion only
after _srcDomXML is inserted into migration parameters.  _srcDomXML
may not have guestAgentAPIVersion metadata in such a case and since
the destination reads metadata from _srcDomXML it doesn't get the
guest API version.

This patch moves update_guest_agent_api_version() call before making
_srcDomXML, ensuring the guestAgentAPIVersion is present in metadata
read by the destination.

Change-Id: I2476ca1f72e2992f6eb20d34316569f120088e00
Bug-Url: https://bugzilla.redhat.com/1788783
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>