vm: Log block job events
When block job (e.g. blockCommit) completes or fails, libvirt sends an
event. We used to log a debug level message showing the libvirt status
and job type enum values:

    unhandled libvirt event (event_name=8, args=("/path", 1, 8))

This message is not very helpful when debug mode is enabled and is not
available by default since debug mode is not enabled by default.

Add VM.onBlockJobEvent(), logging a clear message when a job completes:

    INFO ... Block job COMMIT for drive /path has completed

And error message if job has failed:

    ERROR ... Block job COMMIT for drive /path has failed

We handle also other events such as "ready" and "canceled" and
unexpected job status and type.

We may want to use this event later for optimizing waiting for
completion.

Bug-Url: https://bugzilla.redhat.com/1796415
Change-Id: Ifeb5db52f15b002126b0024c89a10a7df1106b96
Signed-off-by: Nir Soffer <nsoffer@redhat.com>