vm: name live merge cleanup thread based on job ID
Currently we name the thread which does live merge cleanup according
to VM ID for this live merge runs. Thread name is used in the logs
and makes it hard to follow the live merge in the logs when there run
several live merge jobs on the VM. Use job ID for the thread name
instead. The job ID is ID of of the job that does the live merge and
for which cleanup thread is started.

Change-Id: I6cfd9583ed6faa6fbd282244f4b4c669b1e8f480
Bug-Url: https://bugzilla.redhat.com/1796415
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>