qga: rewrite poller
Huge rewrite of the poller. The original version has several design
flaws. It took too long for the initial polling of the started VM to
happen, meaning the VM could spend several minutes without any stats
reported to the engine. This is not something that could be easily fixed
with the old code. Another problem was caused by the threading. The fact
that the queries were grouped by qemu-ga calls caused congestions on a
VM. One call on a VM caused calls from other threads to fail for the
same VM.

All the various Operations were merged into single Operation that is
scheduled often (every 5 seconds in default configuration). All the
qemu-ga commands for all VMs are queried from this operation.

Change-Id: I2e1baf9a122dd2fc1e669c88481c66e1117b4e24
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1680398