Core: Missing readinto function on StreamAdapter
When performing a KVM Import, the importing within the
ovirt-imageio module fails due to a missing function
within the StreamAdapter of the kvm2ovirt module. This
is a regression that breaks the kvm external import.

Added the readinto function to the StreamAdapter in order
to avoid the exception so that the Stream Adapter can
successfully read the proper data.

Change-Id: I970e1d3e596c5af9f03081368e03aebe405b63ee
Bug-Url: https://bugzilla.redhat.com/1798175
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>