qga: return default dict with capabilities
We used to return None if the VM was not yet checked or when the ID was
invalid. But the caller could not reliably distinguish the situations
making the behavior pretty useless. It also complicates some conditions
unnecessarily. Let's return a dictionary in all cases (even for invalid
VM IDs). The caller should still check the version for None to see
whether qemu-ga was reached or not yet.

Change-Id: I46b6d5213a1a598ee9db62a82551d928a9f844fe
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>