hsm: add a wait after connection to iscsi storage domain
Add a 5 seconds (by default)[1] sleep after connecting to an iSCSI storage
domain. This will help alleviate a race where it takes significantly
longer for block devices to be visible for the host after it is logged into
the device.

TODO: A more comprehnsive solution is required here. Once
ovirt-engine will send the list of devices along with the
connectStorageServer command, we could only wait until the devices
we want are visible.

[1] The value can be changed by creating a configuration override file:
    $ cat /etc/vdsm/vdsm.conf.d/iscsi_login_timeout.conf
    [irs]
    iscsi_login_timeout = 2

Change-Id: I4b63e9673a1e1d3447dd7f334453b6e7a5c3a8ea
Bug-Url: https://bugzilla.redhat.com/1807050
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>