virt: async snapshot timeout abort
This patch will let VDSM to abort the async
snapshot operation.
When the snapshot job will start, a start time
value will be saved to the VM's metadata.
A new thread with the purpose to abort the running job
will check the timeout. It will run in both cases of
the job: regular flow or recovery with the
initial timeout.
The default timeout of the snapshot job set
to 30 minutes.
The new thread abort will send abort command to libvirt
job, resulting in ActionStopped VDSM error.

The snapshot metadata is created in the Job itself,
then the shared metadata passed to:
- Snapshot class or Recovery class alternately
- Abort class
The abort class uses only completed/abort variables
from the metadata.
A lock for the Job's module exists to read and write
the metadata to the VM.
The main purpose of the lock is to prevent from
racing on the metadata, in particular the abort
and success status that can be changed by parallel
running threads.

This functional addition is needed since the VM
will be locked by the engine when the snapshot
job will run.
In version 4.4 the job is async, and we want
a finite time to the job, to release the locking.

Change-Id: I4d7107b000e8eac12b6200923fe06ce23f67ec55
Bug-Url: https://bugzilla.redhat.com/1780943
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>