lvm: allow lvm command to overwrite read-only mode
To make sure that lvm metadata are not corrupted by a process running
on a host other than SPM, all lvm commands has to run in read-only mode
on non-SPM hosts. However, there is a few commands (like creating or
destroying VG), which has to run in read-write mode as by the time VG is
created/destoyed, it's attached to a pool and thus these commands are
not executed on SPM. For these commands we have to execute them in
read-write mode. To allow it, add additional parameter to LVMCache.cmd()
which would allow to overwrite LVMCache._read_only value. If the command
specifies the overwrite, it will be used no matter which mode is set
in LVMCache.

Change-Id: I4d168fb0daefc3360fe89325e0efd3c9bbf493c7
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>