lvm: run selected commands in read-write mode
Creating and deleting VGs is not done on SPM but on any host as by that
time SD is detached from the pool. Also extending VG during image merge
can happen on non-SPM hosts. These commands however have to run in
read-write mode. Overwrite read-only mode to read-write when calling
these commands. Also remove all manual switches to read-write mode
in the test, which were a temporal fix to make the tests running.

Change-Id: I549c6f24e8c773604e305137b8d783e4c7930630
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>