storage: Replace get_umask with a thread-safe version
The old code was not thread-safe since the implementation
of "get_umask()" was not a single atomic operation.
It included 2 operations - read & write:
1) Get the old umask (and temporary write a new one instead).
2) Change back to the old umask.
This is a classic example of a non thread-safe flow
when 2 threads might interleave and the result will be
unexpected and inconsistent.
The "get_umask()" is used in "createdir()" methods in
"outOfProcess" and "fileUtils" modules. Both can be executed
by multiple threads, so the protection is required.
The functions is also used in tests, but tests are run
(currently) in a single thread, so tests are not the issue.
The old implementation was more efficient though - an
average of 1000 calls to "get_umask()" in milliseconds:
  Old implementation: 0.0013393709668889642
  New implementation: 0.03459923900663853

Change-Id: I86c3d00543181debef348a842b299ea15c29b591
Signed-off-by: Pavel Bar <pbar@redhat.com>