tests: Limit the number of retries
Update the reproducer script to limit the number of retries so the worst
case when read-only command is repeated we don't create too much delay.
Start with delay of 0.1 seconds, to make sure we don't run the command
too early.

Most failing read-only commands are expected to succeed after one retry,
delaying the command by 0.1 seconds. In the very unlikely worst case we
will have 5 retries with delay of 3.1 seconds.

Here is example run with these settings:

$ python extend.py run-regular \
    --read-only \
    --iterations 100 \
    --extend-delay 10 \
    -p 8000 \
    host5 \
    test \
    /dev/mapper/36001405271fe76b24b542bf858aaeef7 \
    /dev/mapper/360014051ce5179112ae4fb98e72d9ba9 \
    2>run-regular.log

$ python extend.py log-stats run-regular.log
{
    "activating": 5000,
    "creating": 5000,
    "deactivating": 4999,
    "extend-rate": 4.684750527055517,
    "extending": 99996,
    "max-retry": 3,
    "read-only": 109995,
    "refreshing": 99996,
    "removing": 4999,
    "retries": 1374,
    "retry 1": 1339,
    "retry 2": 32,
    "retry 3": 3,
    "retry-rate": 0.012491476885312968,
    "total-time": 21345,
    "warnings": 3766
}

Change-Id: I51bba1a50fafdbcdf5806d8b61b96e0b62de73e9
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>