lvm: Limit number of retries for read-only commands
Limit the number of retries so the worst case when read-only command is
repeated we don't create too much delay. Start with delay of 0.1
seconds, to make sure we don't run the command too early.

Most failing read-only commands are expected to succeed after one retry,
delaying the command by 0.1 seconds. In the very unlikely worst case we
will have 4 retries with delay of 1.5 seconds.

Some commands like lvm.getVG(name), called when creating a new storage
domain will always fail because the VG does not exist. When using
read-only mode we retry such commands because we don't have a good way
to detect failure because of inconsistent metadata, or missing VG.

Before this change, this command was retried 10 times causing a delay of
10 seconds. With this change the command will be retried 4 times,
causing a delay of 1.5 seconds.

Change-Id: I8d42de5f2a444319afe053579a98116ef1714ecf
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>