lvm: Always remove stale VGs and PVs
Previously we removed stale VGs and PVs only when running "vgs" or "pvs"
without specifying the VG or PV names. This was correct since when the
names are specified, a missing VG or PV cause the LVM command to fail,
which drops all VGs or PVs, marking all cached items as Unreadable.

We want to start using --select so LVM commands will not fail when a VG
or PV are missing. In this case we have to remove the missing items from
the cache.

Change-Id: Iee6fcb2654e23b00ab97db3a008a172c8fa66972
ug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>