lvm: Log errors when reloading fail
Previously when LVM command failed, we did not know if this is a real
error (cannot read existing VG) or an expected condition (checking if a
VG exists when creating new storage domain).

Now LVM command will never fail in this case, so any failure is a real
error and must be logged as an error.

Change-Id: Ia0eacdacd289b7af50c40b3464e0d51a47d98a8b
ug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>