lvm: Add temporary _on_rhel() helper to decide on pvs read-write
Fixed lvm2 version for BZ #1809660 would make the read-write override
for pvs command unrequired.  We add a helper for now to detect if we
are running on a RHEL host which should have the appropriate lvm2 rpm.

Once lvm2 rpm would support all target platforms we can remove this
helper and its usages.

Change-Id: If204bdb0d15b0c9ebdfc50490d734c9753c7d3b5
Related-To: https://bugzilla.redhat.com/1809660
Signed-off-by: Amit Bawer <abawer@redhat.com>