lvm: Fix lvm.invalidateVG()
Instead of reloading the VG to find the PVs, iterate over the PVs in the
cache and find all PVs which are part of the VG. This is the same way we
invalidate LVs.

The issue was introduced in when the invalidatePVs option was added:

commit 2f494ab1fdeba3aefcff1ccada47cc6b02f9eb96
Author: Liron Aravot <laravot@redhat.com>
Date:   Sun Jan 22 19:46:15 2017 +0200

    getVGInfo - updated pv information

The only flow using invalidatePVs=True is LVMVolumeGroup.getInfo().

Change-Id: I827dcb0c38e4e25dc114e0e6afdff3125f2d007a
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>