lvm: Fix handling of vgs with empty output
Before moving to --select, vgs with single VG name would fail if the VG
does not exist, marking the VG as Unreadable. This was kind of a bug but
we could not detect that the VG is missing, since lvm return the same
generic error code for all errors.

Now that we use --select, the command does not fail, but we get empty
output. In this case we return the current cache contents, possibly
returning a stale data. Because we never call _reloadvgs() if the VG was
not stale, the VG is mark as Stub. A Stub and Unreadable are basically
the same, but really want to remove non-existing items from the cache.

Fixed by removing the check for empty output, so we always remove stale
vgs. This fixes getVG() so it does not return stubs for missing VG after
reload.

We still process partial output from vgs, so the cache is updated with
the data we have.

Change-Id: I779b80a5d32d6a1ed9748907d2039b15ecfdc90d
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>