lvm: Turn stalelv boolean indication into freshlv set
Use a freshlv set indicating that all lvs are cache fresh for a vg,
when vgName is part of the set.

This would avoid one call for reloadlvs(vgName="vg1", lvNames=None)
turning the old boolean indication of stalelv as False for any
decision to follow for a cache reload in getLV(vgName="vg2", lvName=None)
and wrongfully skipping the cache reload for vg2.

Change-Id: I87f4fdfa9eae0e867b4aa7cc74e17524b0e388fd
Signed-off-by: Amit Bawer <abawer@redhat.com>