tool: Log stdout and stderr as str in exceptions
'ServiceError' class takes stdout and stderr byte streams from arbitrary
commands as its initialization arguments. In the string-conversion
method these are incorrectly treated as strings. Since we cannot
be sure about the contents of the stdout and stderr streams it's better
to log them as string representation of bytes objects that try to decode
them to string.

Change-Id: Ia3f35bfdd824922ac2793805a2f42e0d787c6bed
Bug-Url: https://bugzilla.redhat.com/1813961
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>