net tests: Fix invalid bridge option write
Using nmstate will raise a setupNetwork error if a non existant
option is attempted to be written.

Currently, as the network scripts are still used, adding an invalid
option will raise an AssertionError.
Setting up the network succeeds, since ifup is not able to write
the option, but does not fail as well. The assertion of the
running config vs the requested config is the one to eventually
fail, since the original option was not written in the first place.

Although nmstate is on the brink of replacing ifcfg scripts,
tracking this behavior of network-scripts is a good idea.

Change-Id: Ieb0d2cf2903af99f211f3ec53b93efc5e9cb0388
Signed-off-by: Bell Levin <blevin@redhat.com>