backup: allow performing incremental backup operation
Remove prevention from performing an incremental backup operation and
add 'incremental' attribute to domainbackup XML to mark the backup as
incremental backup.

Incremental backup is not fully supported at this stage, the operation
is valid as long as the VM still running. When the VM will go down or
crash all the backup (and checkpoints) chain cannot be used anymore
and a full backup needs to be taken to start a new chain.

Incremental backup will be fully supported when the API between the
Engine and VDSM for redefining the checkpoints when the VM starts will
be implemented.

Change-Id: I3ff47b78377ae50a4416253c393bcd7072a934e4
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>