betterAsyncore: Proper handling of SSL_ERROR_WANT_WRITE
This patch provides a fix for scenario when there is a 'huge'
Host.getCapabilities response over SSL.
In testing scenario it was only ~200kb to trigger this error.
Previously, only part of the data used to be written
over the socket just before connection was closed because of unhandled
ssl.SSLWantWriteError

The issue has been described at [1] and [2]

Additionally, order of exception handlers has been corrected so that
'sslUtils.SSLError' won't get swallowed. 'OSError' is superclass of
'sslUtils.SSLError' and 'socket.error' is an alias for 'OSError'

$ python3
>>> import socket
>>> assert socket.error is OSError

[1] https://bugs.python.org/issue33307
[2] https://docs.python.org/3/library/ssl.html#ssl-nonblocking

Change-Id: I47b306c9cbdb819e580a96034d9455fac432fe1e
Bug-Url: https://bugzilla.redhat.com/1766193
Signed-off-by: Artur Socha <asocha@redhat.com>