virt: add safety check on md
When cleaning the VM's metadata from
the snapshot job, there is a possibility that
the snapshot data already been cleaned up by
another flow and thread. Then, on a specific bad
timing, it will raise KeyError.

If the metadata is missing, that means the job is
done and cleaned up. No other usage is expected for it.
Therefore, catching it as a KeyError and do nothing is
good handling for this.

Change-Id: I49984262b7e34c8d5457a0a2fa305840457434af
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>