systemd: Do not enable vdsmd after installation
Straight after installation vdsm is not really able to function
properly - it needs at least a 'vdsm-tool configure' run to function
properly.

Currently, when installing vdsm, we put a '85-vdsmd.preset' file
inside systemd's preset directory that causes vdsm to be enabled
by default - this patch removes this file. This will cause
the distro-provided '99-default-disable.preset' preset to be in effect,
which disables everything by default.

Change-Id: I2a69948b55925b77798428648148015f55a72371
Bug-Url: https://bugzilla.redhat.com/1818554
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>