logging: Set UTF-8 encoding for file loggers
If there is a log message containing non-ASCII characters then the log
message is silently dropped.  An example of this problem is a VM with
non-ASCII characters in its name.  Information such as the domain XML
or errors citing the VM name is not logged.

Let's pass `encoding' argument to WatchedFileHandler to fix that.

Change-Id: Id860aae078a8138d2168fb898733306206eacd56
Bug-Url: https://bugzilla.redhat.com/1795206
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>