virt: Fix encoding problems in the libvirt hook
Migrations of VMs with non-ASCII characters in their names fail in the
libvirt migration hook with errors such as:

  xml.etree.ElementTree.ParseError: not well-formed (invalid token): line 2, column 15

The problem is that libvirt sets LC_ALL=C before calling the hook and
that affects sys.stdin and sys.stdout.  Changing the locale later, in
the Python script, doesn't help.  Then Python XML processing assumes
ASCII input and complains on non-ASCII characters, possibly unless an
XML declaration with an explicitly specified encoding is
present (which is not the case with libvirt domain XMLs).

Handling it on a Python level is quite tricky, there are at least the
following problems:

- sys.stdin must be converted to an UTF-8 stream using
  io.TextIOWrapper to get the input encoding set correctly.
- The resulting transformed tree fails in its `write' method due to
  mixture of bytes and strings within it.  The tree must be exported
  using ElementTree.tostring.
- sys.stdout must be tricked by using io.TextIOWrapper too to avoid
  complaints about non-ASCII characters.

Rather than adding the complexity to the Python script, let's simply
override LC_ALL and set it to C.UTF-8 in the script wrapper.  This
shouldn't cause any problems and it makes Python XML processing happy
when using non-ASCII characters.

Change-Id: Ibaa0039800f51a51fef6c3c899ff469a08ce48f9
Bug-Url: https://bugzilla.redhat.com/1795206
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>