net: Add netlink monitor active slaves of a bond
The already existing netlink monitor generates an event which
contains data from the libnl cached object model, triggered
by the receiving of a net netlink message.

The change of the active slave of a mode 1 bond is indicated
by the IFLA_EVENT_BONDING_FAILOVER, which is not included in
libnl's object model.

For this reason, an additional monitor is added, which checks
the netlink message for IFLA_EVENTs.

Change-Id: I3374a4d7d1e08d47d7e49a0d34f7b419c27430a0
Bug-Url: https://bugzilla.redhat.com/1671876
Bug-Url: https://bugzilla.redhat.com/1801794
Signed-off-by: Dominik Holler <dholler@redhat.com>