fakedomainadapter.py: add input_checkpoint_xml member
Add input_checkpoint_xml to store the created checkpoint_xml while
FakeDomainAdapter.backupBegin() is called.

It will allow testing that checkpoint_xml was created as expected in
the test itself and testing the checkpoint XML in a real flow behavior,
so now the specific test for checkpoint XML is redundant.

Change-Id: I83d49f63403aa43933bc1735458bc586a91c7685
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>