gluster: Keep seclabel in network GlusterFS drives
Engine adds seclabel's to those drives.  However on disk hot unplug,
the drive XML is parsed and then built again, while not inserting the
seclabel again for any network drive according to the logic in
storage._getSourceXML.

Let's fix that by adding a check for network GlusterFS drives to
storage._getSourceXML stating that we need seclabel's for them.

Change-Id: Ia121980b4f9fbe36fc672bd47a54140c1789663a
Bug-Url: https://bugzilla.redhat.com/1824805
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>