net: Add mocking of bond defaults for unit tests
Mocking the bonding files creates an independent
environment for unit testing, not to depend on
'/var/run/vdsm'.

By adding the unit tests to use this fixture, it is clear
that all tests are using the bonding defaults generation.
The singular fixtures were removed from the conftest.py
of the specific test folder, and autouse was enable in the
network tests root conftest.py .

Change-Id: I097780460edcc419d6e37776b06915c9f74a9c95
Signed-off-by: Bell Levin <blevin@redhat.com>