API: add new list_checkpoints() API call
New VM.list_checkpoints() API call will be used to fetch the IDs of the
checkpoints that are defined on the given VM ID.

The list of checkpoints IDs that are defined on the VM will allow to the
engine to decide if it should redefine missing checkpoints before
starting an incremental backup.

For example:
vdsm-client VM list_checkpoints vmID=xxx-yyy

{
  'vmID': vm_id
}

The response will be a list of checkpoint IDs that are defined on the VM
ordered from the oldest checkpoint to the leaf.

[
  "checkpoint-id-1",
  "checkpoint-id-2"
]

Change-Id: I3a7e8964d40e36d7827b2804a870d15bffc0b5d8
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>