lvm: Enable read-only pvs command on CentOS
lvm2-2.02.186-7.el7_8.1 is available since 2020-04-08, so we can enable
the fix for bug 1809660 on CentOS.

Unfortunately this fix is not available for ppc64le, so we must keep
using pvs in read-write mode there. The _on_rhel() check was converted
to _on_x86_64() since we don't have this fix on any other architecture
and we are unlikely to get a backport for 7.6.

Change-Id: Ib23783426a73a8a4cef779a8bcaf2613d3cd0197
Bug-Url: https://bugzilla.redhat.com/1829348
Bug-Url: https://bugzilla.redhat.com/1829597
Signed-off-by: Nir Soffer <nsoffer@redhat.com>