network: dhcp_monitor: Fix unix socket handling
The monitor assumed that socket is always removed during shutdown, which
is not valid assumption. For example vdsm may be killed before shutting
down the monitor, or after shutting down the monitor but before calling
os.remove().

This results in this unrecoverable error when starting vdsm:

2020-04-30 13:01:05,097-0400 ERROR (MainThread) [vds] Exception raised (vdsmd:164)
Traceback (most recent call last):
  File "/usr/lib/python3.6/site-packages/vdsm/vdsmd.py", line 162, in run
    serve_clients(log)
  File "/usr/lib/python3.6/site-packages/vdsm/vdsmd.py", line 115, in serve_clients
    init_unprivileged_network_components(cif, supervdsm.getProxy())
  File "/usr/lib/python3.6/site-packages/vdsm/network/initializer.py", line 46, in init_unprivileged_network_components
    dhcp_monitor.initialize_monitor(cif, net_api)
  File "/usr/lib/python3.6/site-packages/vdsm/network/dhcp_monitor.py", line 151, in initialize_monitor
    raise e
  File "/usr/lib/python3.6/site-packages/vdsm/network/dhcp_monitor.py", line 145, in initialize_monitor
    monitor = Monitor.instance()
  File "/usr/lib/python3.6/site-packages/vdsm/network/dhcp_monitor.py", line 94, in instance
    _monitor_instance = Monitor(**kwargs)
  File "/usr/lib/python3.6/site-packages/vdsm/network/dhcp_monitor.py", line 78, in __init__
    socket_path, _MonitorHandler
  File "/usr/lib64/python3.6/socketserver.py", line 456, in __init__
    self.server_bind()
  File "/usr/lib64/python3.6/socketserver.py", line 470, in server_bind
    self.socket.bind(self.server_address)
OSError: [Errno 98] Address already in use

Vdsm ends in endless reboot loop. The only way to recover is to remove
the leftover socket.

The only way to ensure that you can bind to a unix socket is to remove
the socket before trying to bind. Add _remove_socket() helper and call
it before creating the server and after stopping the monitor, but before
shutting it down to make it less likely to leave leftover socket during
shutdown.

When removing the socket, do not warn about missing socket which is
expected condition when starting the server. When logging a warning
after socket removal failed, include the underlying error message
instead of hiding it.

Change-Id: I8674b16582c5522d4a24542f3c382ae61285f897
Signed-off-by: Nir Soffer <nsoffer@redhat.com>