qga: try to detect networks early during the boot
We perform all the qemu-ga queries immediately when we notice that the
agent is running. Often however the network is not yet up and becomes
active only few seconds later. But since we already queried the network
it takes 2 minutes until we attempt to query for changes again. Having
the IP as soon as possible is useful for automation.

What this change actually does is that during boot we check the already
discovered interfaces and IP addresses. If we don't find any addresses
that are not loopback or link-local, explicitly ignoring lo and docker0
interfaces, new query for network interfaces is issued in every poller
iteration.

Change-Id: I9ad96ea7dbe597a61541d154a99da073436238ce
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>