net, api: Set netConfigDirty correctly
netConfigDirty is flag that serves as hint to engine
whether the current network configuration is persisted.
If the flag is set, the network configuration is not
saved and might be lost during reboot.

When commitOnSuccess is used the netConfigDirty was
set as true despite the fact that the configuration
is automatically persisted. Ensure that the flag is
set accordingly.

Change-Id: I53d5d2756364fc4bb16b1a6bebf75d28e6b57856
Bug-Url: https://bugzilla.redhat.com/1798818
Bug-Url: https://bugzilla.redhat.com/1834248
Signed-off-by: Ales Musil <amusil@redhat.com>