numa: Always use int values for online cpu ids
online_cpus in CpuTopology can be set to a list of either string or
integer values, depending on whether libvirt reports NUMA cells or
not.  With NUMA cells reported, a list of str values is retrieved from
libvirt XML, otherwise a list of int values is obtained from
taskset.online_cpus().  In the latter case, a type error is raised
when the list of integers is joined as strings while onlineCpus value
is created in caps.get().

Let's unify the types in CpuTopology to integers, to be consistent
with CPU ids elsewhere; and let's fix caps.get(), the only place that
uses CpuTopology.online_cpus, to convert them to strings.

Change-Id: I9848f7c6bb4b11084f54b8b9a1cd63d4bb1dcd01
Bug-Url: https://bugzilla.redhat.com/1834873
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>