qga: add context for setting libvirt timeout
With libvirt 5.10 we get the ability to configure timeout for qemu-ga
commands. This is important for stats collection because one rogue guest
can block collection for other VMs. The complication with the API is
that it is not per-command but sets the timeout globaly for all commands
of the VM. But for the commands not related to stats collection we still
want to rely on the defaults.

Here we introduce a context for setting the timeout on commands. It
internally serializes the request by using a lock mechanism.
The timeout is only valid for the duration of the context and is reset
to `default` when context ends.

Change-Id: I7f2d8a8194b9b0e6bc7088b9765be752a25fd303
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>