virt: snapshot failure handling
Using the jobs framework, it accepts only unhandled exceptions to
determaine the result of the job. When we returned error response code
it marked the job as 'done'. Therefore, the engine would get the job as
succeed.

This patch changes the exception handling in snapshot job from returning
error response code to raising unhandled exception that will be caught
in jobs framework and mark the job as a failure.

Change-Id: I92bed3a490d43df2adc68a33446e9d6523c8d318
Bug-Url: https://bugzilla.redhat.com/1835096
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>