net, nmstate: Send all dhcp info in event
The dispatcher dhcp_monitor was sending only
information that were available however vdsmd was
expecting everything. This was not causing any issues
with ipv4 because we had everything we needed.

Ipv6 on the other hand did miss some information.
Sending none instead of missing field will prevent
exceptions thrown because of that.

This mechanism was heavily relied on by functional
tests. This fix speeds their execution by up to 15%
depending on machine.

Change-Id: I9ee74ac7f7b47d27306c47ef3ed98a926592f0f5
Signed-off-by: Ales Musil <amusil@redhat.com>