hooks: vhostmd: Handle virtio channels
Vhostmd 1.1 introduces new way for the VMs to communicate with 'vhostmd'
daemon running on the host - a virtio serial channel. This patch adds
the support for using this transport (on top of the existing block
device method, which is still relevant) to the 'vhostmd' hook
along with some tests.

Insertion of the block device and the virtio serial channel is decided
dynamically and depends on the configuration of 'vhostmd'. Both can
(and will, if the config says so) be enabled simultanously.

[1] https://github.com/vhostmd/vhostmd/blob/52b2dbf5c7136f87b1e6a6f4a10c363779bd1fb4/README

Change-Id: Id44ee43d3ad3deb08f28d542d239896c035b1da2
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>