backup.py: implement delete_checkpoints
Implement delete_checkpoints verb.

Deleting checkpoint is needed by the Engine in the following
cases:
  - For reducing the number of known checkpoints for a VM,
    delete X oldest checkpoints to keep number of checkpoints small.

  - When the Engine and libvirt cannot be synced, delete all checkpoints.

The given checkpoints IDs is a list of checkpoint IDs reported by
libvirt to the Engine as checkpoints that are defined on the VM, the
list is arranged from the oldest checkpoint ID to the newest checkpoint
ID.

In case of failure to remove a checkpoint from the chain, the operation
will stop and will not continue to remove the following checkpoints on
the list. In that case, the details of the error will be attached to the
response:

{
  'checkpoint_ids': [removed_checkpoint_1, removed_checkpoint_2, ... ],
  'error': {
    'code': err_code,
    'message': "err_msg"
  }
}

Change-Id: I949c629d54cf3d87cea2ec3bc06551db081f6011
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>