fix JSON-RPC error when fencing fails
This patch fixes the fence method to get rid of execCmd to run the
fencing script and use the start method instead since execCmd is
deprecated and python3 execCmd will return the error and output as binary arrays.

In addition the output/error may be binary and should be converted to
utf-8 before the result is returned to the caller.

Change-Id: I0e98885845500600605507707060dc04618a919c
Bug-Url: https://bugzilla.redhat.com/1810974
Signed-off-by: Eli Mesika <emesika@redhat.com>